update menu / admin / localization

+++ src/Controller/UserController.php
@@ -9,14 +9,48 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
+
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use App\Service\RandomString;
 
 /**
- * @Route("/user")
+ * @Route("/admin/user")
  */
 class UserController extends AbstractController
 {
     /**
-     * @Route("/", name="user_index", methods="GET")
+    * @var ValidatorInterface
+    */
+    private $validator;
+
+    /**
+    * @var RandomString
+    */
+    private $random_string;
+
+    /**
+    * @var Security
+    */
+    private $security;
+
+    /**
+    * @var User
+    */
+    private $user;
+
+    public function __construct(ValidatorInterface $validator, RandomString $random_string, Security $security)
+    {
+        $this->validator = $validator;
+        $this->random_string = $random_string;
+        $this->security = $security;
+        $this->user = $this->security->getUser();
+    }
+
+    /**
+     * @Route("/", name="admin_user_index", methods="GET")
      */
     public function index(): Response
     {
@@ -25,30 +59,36 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="user_new", methods="GET|POST")
+     * @Route("/new", name="admin_user_new", methods="GET|POST")
      */
-    public function new(Request $request): Response
+    public function new(Request $request, UserPasswordEncoderInterface $passwordEncoder): Response
     {
         $user = new User();
         $form = $this->createForm(User1Type::class, $user);
         $form->handleRequest($request);
+        
+        $errors = $this->validator->validate($user);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            //correctly encode password
+            $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
+
             $em = $this->getDoctrine()->getManager();
             $em->persist($user);
             $em->flush();
 
-            return $this->redirectToRoute('user_index');
+            return $this->redirectToRoute('admin_user_index');
         }
 
         return $this->render('user/new.html.twig', [
             'user' => $user,
+            'errors' => $errors,
             'form' => $form->createView(),
         ]);
     }
 
     /**
-     * @Route("/{id}", name="user_show", methods="GET")
+     * @Route("/{id}", name="admin_user_show", methods="GET")
      */
     public function show(User $user): Response
     {
@@ -56,36 +96,51 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/edit", name="user_edit", methods="GET|POST")
+     * @Route("/{id}/edit", name="admin_user_edit", methods="GET|POST")
      */
-    public function edit(Request $request, User $user): Response
+    public function edit(Request $request, UserPasswordEncoderInterface $passwordEncoder, User $user): Response
     {
         $form = $this->createForm(User1Type::class, $user);
         $form->handleRequest($request);
 
+        $errors = $this->validator->validate($user);
+
         if ($form->isSubmitted() && $form->isValid()) {
+            //correctly encode password
+            $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
+
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('user_edit', ['id' => $user->getId()]);
+            return $this->redirectToRoute('admin_user_edit', ['id' => $user->getId()]);
         }
 
         return $this->render('user/edit.html.twig', [
             'user' => $user,
+            'errors' => $errors,
             'form' => $form->createView(),
         ]);
     }
 
     /**
-     * @Route("/{id}", name="user_delete", methods="DELETE")
+     * @Route("/{id}", name="admin_user_delete", methods="DELETE")
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($user);
-            $em->flush();
+        if( $this->user->hasRole('ROLE_ADMIN') )
+        {
+            if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($user);
+                $em->flush();
+            }
+        }
+        else{
+            $this->addFlash(
+                'notice',
+                'You can\'t delete a user!'
+            );
         }
 
-        return $this->redirectToRoute('user_index');
+        return $this->redirectToRoute('admin_user_index');
     }
 }
